Deploy UI builds for pull requests.
Pull request builds do not have access to secrets for security reasons,
so cannot deploy directly. Instead we create and upload the build in the
build-pull-request workflow and deploy to temporary preview URL's in the
deploy-pull-request workflow. The deploy workflow is triggered in
response to the build workflow, and runs against the repo's main branch
. It has access to secrets and can run the deploy commands.

The preview origins are separate from production origins, so the builds
cannot exploit XSS. Also, we generally trust the code coming from our
contributors since we trust the contributors. First-time or unsolicited
contributors need to be approved before workflows will run on their pull
requests as well, so there is little chance that untrusted code wouldbe
deployed. Even if it is, it has no access to secrets and is at least as
secure as checking out the code, building, and running locally.