Make sure relationships are only defined once Defining relationships from both ends can lead to inconsistencies and confusion. The update leaves only one definition per relationship and creates properties for accessing related nodes more easily from the non-owning side.
The following nodes have been updated:
- Litigation
- Agency
- Unit
- Civilian
- Complaint
-- and children
- Officer
- StateID

In cases where there were parent/child relationships, the child nodes
have been made the owners of the relationships. This is done because
a child node must have a cardinality of `One`` or `ZeroOrOne` and this
can only be defined on the owning side of a relationship.