Make sure relationships are only defined once (#467)
* Make sure relationships are only defined once
Defining relationships from both ends can lead to inconsistencies and
confusion. The update leaves only one definition per relationship and
creates properties for accessing related nodes more easily from the
non-owning side.

The following nodes have been updated:
- Litigation
- Agency
- Unit
- Civilian
- Complaint
-- and children
- Officer
- StateID

In cases where there were parent/child relationships, the child nodes
have been made the owners of the relationships. This is done because
a child node must have a cardinality of `One`` or `ZeroOrOne` and this
can only be defined on the owning side of a relationship.

* Use `RelQuery` properties instad of direct queries
Creates a `RelQuery` class for querying relationships from the
non-owning side. This mimics some of the more useful elements of the
neomodel `RelationshipManager` without the unnecessary overhead or the
ability to create relationships.

* Add `virtual_relationships` for serializer
The `JSONSerializable` class now supports virtual relationships for
nodes that don't own the relationship manager but still need to return
related nodes as part of their dictionary representation.
The `__virtual_relationships__` attribute will only work with properties
that return `RelQuery` objects.