Convert Citation relationship to a Mixin Class (#508)
* Convert Citation relationship to a mixin
Citation usage and management is needed across multiple models. To
avoid code duplication, we convert the Citation relationship into a
mixin that can be inherited by any model that needs it.

The mixin will encapsulate the relationship definition and any
related methods, including adding citations and determining the primary
source for a node.

* Citation mixin no longer abstract
Standardize primary_source:
  - Use elementId instead of uid
  - Always a property, not a method

* Remove helper functions from `primary_source`